+++ api/src/main/java/com/lovely4k/backend/diary/service/DiaryService.java
@@ -5,12 +5,11 @@ import com.lovely4k.backend.couple.Couple;
 import com.lovely4k.backend.couple.IncreaseTemperatureEvent;
 import com.lovely4k.backend.couple.repository.CoupleRepository;
 import com.lovely4k.backend.diary.Diary;
-import com.lovely4k.backend.diary.DiaryRepositoryAdapter;
+import com.lovely4k.backend.diary.DiaryRepository;
 import com.lovely4k.backend.diary.Photos;
 import com.lovely4k.backend.diary.controller.request.DiaryDeleteRequest;
 import com.lovely4k.backend.diary.service.request.DiaryCreateRequest;
 import com.lovely4k.backend.diary.service.request.DiaryEditRequest;
-import com.lovely4k.backend.diary.service.response.*;
 import com.lovely4k.backend.member.Member;
 import com.lovely4k.backend.member.Sex;
 import com.lovely4k.backend.member.repository.MemberRepository;
@@ -21,7 +20,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Slf4j
@@ -31,7 +29,7 @@ import java.util.List;
 public class DiaryService {
 
     private final MemberRepository memberRepository;
-    private final DiaryRepositoryAdapter diaryRepositoryAdapter;
+    private final DiaryRepository diaryRepository;
     private final CoupleRepository coupleRepository;
     private final DiaryImageUploader diaryImageUploader;
 
@@ -43,7 +41,7 @@ public class DiaryService {
         List<String> uploadedImageUrls = diaryImageUploader.uploadImages(multipartFileList);
         Diary diary = diaryCreateRequest.toEntity(member);
         diary.addPhoto(Photos.create(uploadedImageUrls));
-        Diary savedDiary = diaryRepositoryAdapter.save(diary);
+        Diary savedDiary = diaryRepository.save(diary);
 
         Events.raise(new IncreaseTemperatureEvent(member.getCoupleId()));
         return savedDiary.getId();
@@ -64,17 +62,9 @@ public class DiaryService {
     }
 
     private Diary validateDiaryId(Long diaryId) {
-        return diaryRepositoryAdapter.findById(diaryId).orElseThrow(
+        return diaryRepository.findById(diaryId).orElseThrow(
                 () -> new EntityNotFoundException("invalid diary id")
         );
-    }
-
-    public DiaryListInGridResponse findDiaryListInGrid(BigDecimal rLatitude, BigDecimal rLongitude, BigDecimal lLatitude, BigDecimal lLongitude, Long coupleId) {
-        List<Diary> diaryList = diaryRepositoryAdapter.findDiaryList(rLatitude, rLongitude, lLatitude, lLongitude, coupleId);
-
-        return new DiaryListInGridResponse(diaryList.stream().map(
-            DiaryGridResponse::from
-        ).toList());
     }
 
     @Transactional
@@ -115,7 +105,7 @@ public class DiaryService {
     public void deleteDiary(Long diaryId, Long coupleId) {
         Diary diary = validateDiaryId(diaryId);
         diary.checkAuthority(coupleId);
-        diaryRepositoryAdapter.delete(diary);
+        diaryRepository.delete(diary);
     }
 
     @Transactional
@@ -128,6 +118,6 @@ public class DiaryService {
             }
         ).toList();
 
-        diaryRepositoryAdapter.deleteAll(diaries);
+        diaryRepository.deleteAll(diaries);
     }
 }
