+++ api/src/test/java/com/lovely4k/backend/diary/service/DiaryQueryServiceTest.java
@@ -6,10 +6,7 @@ import com.lovely4k.backend.couple.service.response.InvitationCodeCreateResponse
 import com.lovely4k.backend.diary.Diary;
 import com.lovely4k.backend.diary.DiaryRepository;
 import com.lovely4k.backend.diary.Photos;
-import com.lovely4k.backend.diary.service.response.WebDiaryDetailResponse;
-import com.lovely4k.backend.diary.service.response.WebDiaryListByMarkerResponse;
-import com.lovely4k.backend.diary.service.response.WebDiaryListResponse;
-import com.lovely4k.backend.diary.service.response.WebDiaryMarkerResponse;
+import com.lovely4k.backend.diary.service.response.*;
 import com.lovely4k.backend.location.Category;
 import com.lovely4k.backend.location.Location;
 import com.lovely4k.backend.location.LocationRepository;
@@ -219,6 +216,77 @@ class DiaryQueryServiceTest extends IntegrationTestSupport {
 
         // then
         assertThat(webDiaryListByMarkerResponse).isNull();
+    }
+
+    @DisplayName("findDiaryListInGrid를 통해 위치 기반 다이어리 목록을 조회할 수 있다.")
+    @Test
+    void findDiaryListInGrid() {
+        // given
+        Diary diary1 = Diary.builder()
+            .coupleId(1L)
+            .location(Location.create(1L, "서울시", "서울역", BigDecimal.valueOf(37.5563), BigDecimal.valueOf(126.9723), Category.ETC))
+            .boyText("hello")
+            .girlText("hi")
+            .score(4)
+            .photos(Photos.builder().firstImage("test-image1").build())
+            .datingDay(LocalDate.of(2023, 10, 20))
+            .build();
+
+        Diary diary2 = Diary.builder()
+            .coupleId(1L)
+            .location(Location.create(1098L, "부산시", "부산역", BigDecimal.valueOf(35.1151), BigDecimal.valueOf(129.0422), Category.ETC))
+            .boyText("hello")
+            .girlText("hi")
+            .score(3)
+            .photos(Photos.builder().firstImage("test-image1").build())
+            .datingDay(LocalDate.of(2023, 10, 20))
+            .build();
+
+        diaryRepository.saveAll(List.of(diary1, diary2));
+
+        // when
+        DiaryListInGridResponse result = diaryQueryService.findDiaryListInGrid(BigDecimal.valueOf(38), BigDecimal.valueOf(127), BigDecimal.valueOf(36), BigDecimal.valueOf(126), 1L);
+
+        // then
+        assertAll(
+            () -> assertThat(result.diaries()).hasSize(1),
+            () -> assertThat(result.diaries().get(0).diaryId()).isEqualTo(diary1.getId())
+        );
+    }
+
+    @DisplayName("만족하는 결과가 없을경우 findDiaryListInGrid의 diaries는 Empty를 반환한다. ")
+    @Test
+    void findDiaryListInGrid_Empty() {
+        // given
+        Diary diary1 = Diary.builder()
+            .coupleId(1L)
+            .location(Location.create(1L, "서울시", "서울역", BigDecimal.valueOf(37.5563), BigDecimal.valueOf(126.9723), Category.ETC))
+            .boyText("hello")
+            .girlText("hi")
+            .score(4)
+            .photos(Photos.builder().firstImage("test-image1").build())
+            .datingDay(LocalDate.of(2023, 10, 20))
+            .build();
+
+        Diary diary2 = Diary.builder()
+            .coupleId(1L)
+            .location(Location.create(1098L, "부산시", "부산역", BigDecimal.valueOf(35.1151), BigDecimal.valueOf(129.0422), Category.ETC))
+            .boyText("hello")
+            .girlText("hi")
+            .score(3)
+            .photos(Photos.builder().firstImage("test-image1").build())
+            .datingDay(LocalDate.of(2023, 10, 20))
+            .build();
+
+        diaryRepository.saveAll(List.of(diary1, diary2));
+
+        // when
+        DiaryListInGridResponse result = diaryQueryService.findDiaryListInGrid(BigDecimal.valueOf(35), BigDecimal.valueOf(127), BigDecimal.valueOf(34), BigDecimal.valueOf(126), 1L);
+
+        // then
+        assertAll(
+            () -> assertThat(result.diaries()).isEmpty()
+        );
     }
 
     private static Diary buildDiary(Location location, long coupleId) {
