+++ model/src/main/java/com/lovely4k/backend/diary/DiaryQueryRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RequiredArgsConstructor
@@ -29,4 +30,7 @@ public class DiaryQueryRepository {
         return diaryRepository.findByMarker(kakaoMapId, coupleId);
     }
 
+    public List<Diary> findDiaryListInGrid(BigDecimal rLatitude, BigDecimal rLongitude, BigDecimal lLatitude, BigDecimal lLongitude, Long coupleId) {
+        return diaryRepository.findInGrid(rLatitude, rLongitude, lLatitude, lLongitude, coupleId);
+    }
 }
